requirements.txt
streamlit
requests
beautifulsoup4
deep-translator
pandas
google-generativeai